Refactor to introduce a data format base class.
